Move `make_command_transaction`
Create a families/command/workload directory and move the
`make_command_transaction` method to its own module in the workload
directory. This module is behind the experimental feature
"family-command-workload".
